requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pandas==2.1.3
alembic==1.12.1
python-dotenv==1.0.0
pydantic==2.5.1
